[S08P31A603-433] Feat: 기록 즐겨찾기 수정 API 구현

+++ backend/hearo/src/main/java/com/ssafy/hearo/domain/conversation/entity/Record.java
@@ -66,8 +66,8 @@ public class Record {
         this.title = title;
     }
 
-    public void modifyFavorite(byte isFavorite) {
-        this.isFavorite = isFavorite;
+    public void modifyFavorite(long isFavorite) {
+        this.isFavorite = (byte)isFavorite;
     }
 
     public void delete() {
